% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidNucleobasesInDNA.R
\name{checkValidNucleobasesInDNA}
\alias{checkValidNucleobasesInDNA}
\title{Checks if given DNA sequence is composed with valid characters representing the nucleobases.}
\usage{
checkValidNucleobasesInDNA(sequence)
}
\arguments{
\item{sequence}{A vector of characters in lower or upper case indicating some DNA sequence.}
}
\value{
TRUE if the sequence is valid DNA sequence, otherwise returns FALSE
}
\description{
Checks if given DNA sequence is composed with valid characters representing the nucleobases.
}
\examples{
checkValidNucleobasesInDNA(c("A","C","A","C","A","C","T","A"))
checkValidNucleobasesInDNA(c("a","g","c","a","c","a","c","a"))
}
