% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidNucleobasesInRNA.R
\name{checkValidNucleobasesInRNA}
\alias{checkValidNucleobasesInRNA}
\title{Checks if given RNA sequence is composed with valid characters representing the nucleobases.}
\usage{
checkValidNucleobasesInRNA(sequence)
}
\arguments{
\item{sequence}{A vector of characters in lower or upper case indicating some RNA sequence.}
}
\value{
TRUE if the sequence is valid RNA sequence, otherwise returns FALSE
}
\description{
Checks if given RNA sequence is composed with valid characters representing the nucleobases.
}
\examples{
checkValidNucleobasesInRNA(c("A","C","A","C","A","C","U","A"))
checkValidNucleobasesInRNA(c("a","g","c","a","c","a","c","a"))
}
