% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\title{Finds the complement of nucleobase in DNA or RNA.}
\usage{
complement(nucleobase, mode)
}
\arguments{
\item{nucleobase}{A character in lower or upper case indicating some of the nucleobases in DNA or RNA.}

\item{mode}{String 'dna' or 'rna' indicationg the nucleic acid}
}
\value{
The complement of \code{nucleobase}.
}
\description{
Finds the complement of nucleobase in DNA or RNA.
}
\examples{
complement('A', 'dna')
complement('c', 'dna')
complement('U', 'rna')
complement('g', 'rna')
}
