% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAligmentMatrix.R
\name{createAligmentMatrix}
\alias{createAligmentMatrix}
\title{Creates aligment matrix for given DNA sequences.}
\usage{
createAligmentMatrix(DNA, startPositions, t, l)
}
\arguments{
\item{DNA}{A matrix with DNA sequences in each row.}

\item{startPositions}{A vector with start position for every DNA sequence.}

\item{t}{A number indicating the number of DNA sequences.}

\item{l}{Lenght of the motif.}
}
\value{
Aligment matrix
}
\description{
Creates aligment matrix for given DNA sequences.
}
\examples{
DNA <- matrix(c("A","C","A","C","A","C","G","A","A","C","U","G","A","C","G","C","A","C","G","G","G","C","U","A"), nrow=4, ncol=6)
startPositions <- c(1,2,1,3)
t <- nrow(DNA)
createAligmentMatrix(DNA, startPositions, t, 4)
}
