% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProfileMatrix.R
\name{createProfileMatrix}
\alias{createProfileMatrix}
\title{Creates profile matrix for given aligment matrix.}
\usage{
createProfileMatrix(aligmentMatrix)
}
\arguments{
\item{aligmentMatrix}{A matrix indicating the aligment matrix.}
}
\value{
Profile matrix
}
\description{
Creates profile matrix for given aligment matrix.
}
\examples{
createProfileMatrix(matrix(c("A","C","A","C","A","C","U","A","A","C","U","A"), nrow=4, ncol=3))
}
