% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMotifWithConsensus.R
\name{findMotifWithConsensus}
\alias{findMotifWithConsensus}
\title{Finds score and consensus along with aligment and profile matrices for given DNA sequences and start positions.}
\usage{
findMotifWithConsensus(DNA, startPositions, l)
}
\arguments{
\item{DNA}{A matrix with DNA sequences in each row.}

\item{startPositions}{A vector with start position for every DNA sequence.}

\item{l}{Lenght of the motif.}
}
\value{
List with the aligment matrix, profile matrix, score and consensus.
}
\description{
Finds score and consensus along with aligment and profile matrices for given DNA sequences and start positions.
}
\examples{
DNA <- matrix(c("A","C","A","C","A","C","G","A","A","C","U","G","A","C","G","C","A","C","G","G","G","C","U","A"), nrow=4, ncol=6)
startPositions <- c(1,2,1,3)
findMotifWithConsensus(DNA, startPositions, 4)
}
