% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticCoding.R
\name{geneticCoding}
\alias{geneticCoding}
\title{Codes some DNA codon into aminoacid.}
\usage{
geneticCoding(codon)
}
\arguments{
\item{codon}{A string with tree characters in lower or upper case indicating three nucleotides from DNA codon.}
}
\value{
Character indicating the aminoacid as a result from the genetic code.
}
\description{
Codes some DNA codon into aminoacid.
}
\examples{
geneticCoding("ACA")
geneticCoding("gta")
}
