% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticDecoding.R
\name{geneticDecoding}
\alias{geneticDecoding}
\title{Decodes given aminoacid to DNA codon/s.}
\usage{
geneticDecoding(aminoacid)
}
\arguments{
\item{codon}{A character indicating the aminoacid.}
}
\value{
Vector with the coresponding DNA codons in it.
}
\description{
Decodes given aminoacid to DNA codon/s.
}
\examples{
geneticDecoding("M")
geneticDecoding("S")
geneticDecoding("*")
}
