% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComplementaryStrand.R
\name{getComplementaryStrand}
\alias{getComplementaryStrand}
\title{Finds the complementary strand for given DNA strand}
\usage{
getComplementaryStrand(dna)
}
\arguments{
\item{dna}{A vector of characters in lower or upper case indicating some DNA sequence.}
}
\value{
Vector of characters reperesenting the complementary DNA strand.
}
\description{
Finds the complementary strand for given DNA strand
}
\examples{
getComplementaryStrand(c("A","C","A","C","A","C","T","A"))
getComplementaryStrand(c("a","g","c","a","c","a","c","a"))
}
