% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStartPositionsWithMedianString.R
\name{getStartPositionsWithMedianString}
\alias{getStartPositionsWithMedianString}
\title{Finds score and consensus along with aligment and profile matrices for given DNA sequences and start positions.}
\usage{
getStartPositionsWithMedianString(DNA, consensus, motifLen)
}
\arguments{
\item{DNA}{A matrix with DNA sequences in each row.}

\item{consensus}{A vector with start position for every DNA sequence.}

\item{motifLen}{Lenght of the motif.}
}
\value{
Vector with starting positions of the DNA sequences.
}
\description{
Finds score and consensus along with aligment and profile matrices for given DNA sequences and start positions.
}
\examples{
DNA <- matrix(c("A","C","A","C","A","C","G","A","A","C","U","G","A","C","G","C","A","C","G","G","G","C","U","A"), nrow=4, ncol=6)
consensus <- c("A","C","A","A")
getStartPositionsWithMedianString(DNA, consensus, 4)
}
