% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAligmentNucleotideSequence.R
\name{globalAligmentNucleotideSequence}
\alias{globalAligmentNucleotideSequence}
\title{Finds the optimal global nucleotide sequence aligment with Needleman-Wunsch algorithm with BLOSUM50 as score matrix.}
\usage{
globalAligmentNucleotideSequence(sequence1, sequence2, gapPenalty,
  matchingScore, unmatchingScore)
}
\arguments{
\item{sequence1}{First vector with characters indicating some of DNA nucleotides.}

\item{sequence2}{Second vector with characters indicating some of DNA nucleotides.}

\item{gapPenalty}{A number indicating the gap penalty value.}

\item{matchingScore}{A number indicating the value for matching nucleotides.}

\item{unmatchingScore}{A number indicating the value for unmatching nucleotides.}
}
\value{
List containing the score, the resulting aligned sequences and the matrix of aligment.
}
\description{
Finds the optimal global nucleotide sequence aligment with Needleman-Wunsch algorithm with BLOSUM50 as score matrix.
}
\examples{
globalAligmentNucleotideSequence(c("A","C","A","C","A","C","T","A"), c("A","G","C","A","C","A","C","A"), 1, 2, -1)
}
