% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAligmentProteinSequence.R
\name{globalAligmentProteinSequence}
\alias{globalAligmentProteinSequence}
\title{Finds the optimal global nucleotide sequence aligment with Needleman-Wunsch algorithm with BLOSUM50 as score matrix.}
\usage{
globalAligmentProteinSequence(sequence1, sequence2, gapPenalty, blosumName)
}
\arguments{
\item{sequence1}{First vector with characters indicating some protein sequence.}

\item{sequence2}{Second vector with characters indicating some protein sequence.}

\item{gapPenalty}{A number indicating the gap penalty value.}

\item{blosumName}{Name of the BLOSUM matrix.}
}
\value{
List containing the score, the resulting aligned sequences and the matrix of aligment.
}
\description{
Finds the optimal global nucleotide sequence aligment with Needleman-Wunsch algorithm with BLOSUM50 as score matrix.
}
\examples{
globalAligmentProteinSequence(c("P","A","W","H","E","A","E"), c("H","E","A","G","A","W","G","H","E","E"), 8, "BLOSUM50")
}
