% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammingDistance.R
\name{hammingDistance}
\alias{hammingDistance}
\title{Calculates hamming distance between two DNA sequences with same lenght.}
\usage{
hammingDistance(seq1, seq2)
}
\arguments{
\item{seq1}{A vector with characters indicating first DNA sequence.}

\item{seq2}{A vector with characters indicating second DNA sequence.}
}
\value{
Character indicating the aminoacid as a result from the genetic code.
}
\description{
Calculates hamming distance between two DNA sequences with same lenght.
}
\examples{
hammingDistance(c("A","C","A","A","C","T","C","A","T"), c("A","G","A","G","C","T","C","A","A"))
}
