% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localAligmentNucleotideSequence.R
\name{localAligmentNucleotideSequence}
\alias{localAligmentNucleotideSequence}
\title{Finds the optimal local nucleotide sequence aligment with Smith-Waterman algorithm and BLOSUM50 as score matrix.}
\usage{
localAligmentNucleotideSequence(sequence1, sequence2, gapPenalty, matchingScore,
  unmatchingScore)
}
\arguments{
\item{sequence1}{First vector with characters indicating some of DNA nucleotides.}

\item{sequence2}{Second vector with characters indicating some of DNA nucleotides.}

\item{gapPenalty}{A number indicating the gap penalty value.}

\item{matchingScore}{A number indicating the value for matching nucleotides.}

\item{unmatchingScore}{A number indicating the value for unmatching nucleotides.}
}
\value{
List containing the score and the resulting aligned sequences.
}
\description{
Finds the optimal local nucleotide sequence aligment with Smith-Waterman algorithm and BLOSUM50 as score matrix.
}
\examples{
localAligmentNucleotideSequence(c("A","C","A","C","A","C","T","A"), c("A","G","C","A","C","A","C","A"), 1, 2, -1)
}
