% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAGeneticCoding.R
\name{RNAGeneticCoding}
\alias{RNAGeneticCoding}
\title{Codes some DNA codon into aminoacid.}
\usage{
RNAGeneticCoding(codon)
}
\arguments{
\item{codon}{A string with tree characters in lower or upper case indicating three nucleotides from mRNA codon.}
}
\value{
Character indicating the aminoacid as a result from the genetic code.
}
\description{
Codes some DNA codon into aminoacid.
}
\examples{
RNAGeneticCoding("ACA")
RNAGeneticCoding("gua")
}
