% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAGeneticDecoding.R
\name{RNAGeneticDecoding}
\alias{RNAGeneticDecoding}
\title{Decodes given aminoacid to RNA codon/s.}
\usage{
RNAGeneticDecoding(aminoacid)
}
\arguments{
\item{codon}{A character indicating the aminoacid.}
}
\value{
Vector with the coresponding RNA codons in it.
}
\description{
Decodes given aminoacid to RNA codon/s.
}
\examples{
RNAGeneticDecoding("M")
RNAGeneticDecoding("S")
RNAGeneticDecoding("*")
}
