% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Finds all protein sequences for the given DNA sequence. First the algorithm finds the RNA for the given
DNA sequence. Then it starts from the begining searching for a start codon, and stops when finds stop
codon of end of sequence.When one sequence is found, the algorithm searches for new start codon in the
sequence, after the previous start codon.}
\usage{
translation(sequence)
}
\arguments{
\item{sequence}{A vector of characters in lower or upper case indicating some DNA sequence.}
}
\value{
A list with vectors indicating all possible protein sequences for given DNA.
}
\description{
Finds all protein sequences for the given DNA sequence. First the algorithm finds the RNA for the given
DNA sequence. Then it starts from the begining searching for a start codon, and stops when finds stop
codon of end of sequence.When one sequence is found, the algorithm searches for new start codon in the
sequence, after the previous start codon.
}
\examples{
translation(c("A","T","G","C","A","C","A","T","G","T","A","C","T","G","A","T","A","C"))
}
