% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRNAforDNA.R
\name{findRNAforDNA}
\alias{findRNAforDNA}
\title{Finds the mRNA sequence for given DNA sequence}
\usage{
findRNAforDNA(dna)
}
\arguments{
\item{dna}{A vector of characters in lower or upper case indicating some DNA sequence.}
}
\value{
Vector of characters reperesenting mRNA sequence for given DNA.
}
\description{
Finds the mRNA sequence for given DNA sequence
}
\examples{
findRNAforDNA(c("A","C","A","C","A","C","T","A"))
findRNAforDNA(c("a","g","c","a","c","a","c","a"))
}
